+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 <!-- There should always be "Unreleased" section at the beginning. -->
 
 ## Unreleased
+- Allow setting an edismax as local parameters in query
+  - [**BC**] Add method `useEDisMaxGlobally` to `FulltextInterface`
+- Set `phraseBigramFields`, `boostQuery`, `phraseSlop` only when `eDisMax` is enabled
 
 ## 1.3.0 - 2022-03-29
 - Use `*` as a placeholder for all fields in `EntityApplicator`